clearing the post id from the user model after a post is deleted

+++ src/controllers/feed.ts
@@ -116,6 +116,7 @@ export const getSinglePost = (req: Request, res: Response, next: NextFunction) =
         error.statusCode = 404;
         throw error;
       }
+
       res.status(200).json({
         message: 'Post Fetched.',
         post,
@@ -161,6 +162,12 @@ export const updatePost = (req: Request, res: Response, next: NextFunction) => {
         throw error;
       }
 
+      if (post?.creator?.toString() !== req?.userId) {
+        const error: HttpError = new Error('Not Authorized.');
+        error.statusCode = 403;
+        throw error;
+      }
+
       if (imageUrl !== post?.imageUrl) {
         clearImage(post?.imageUrl);
       }
@@ -196,12 +203,29 @@ export const deletePost = (req: Request, res: Response, next: NextFunction) => {
         error.statusCode = 404;
         throw error;
       }
-      //TODO: check logged in user
+
+      if (post?.creator?.toString() !== req?.userId) {
+        const error: HttpError = new Error('Not Authorized.');
+        error.statusCode = 403;
+        throw error;
+      }
+
       clearImage(post?.imageUrl);
       return Post.findByIdAndRemove(postId);
     })
     .then((result) => {
-      console.log('delete post result', result);
+      return User.findById(req.userId);
+    })
+    .then((user) => {
+      if (!user) {
+        const error: HttpError = new Error('User Not Found!');
+        error.statusCode = 500;
+        throw error;
+      }
+      user.posts?.pull(postId);
+      return user.save();
+    })
+    .then((result) => {
       res.status(200).json({
         message: 'Post deleted successfully',
       });
